--- backend/.eslintrc.js
@@ -1,19 +1,0 @@
-module.exports = {
-    env: {
-        node: true,
-        commonjs: true,
-        es2021: true,
-    },
-    extends: 'eslint:recommended',
-    overrides: [
-        {
-            files: ['.eslintrc.{js,cjs}'],
-            parserOptions: {
-                sourceType: 'script',
-            },
-        },
-    ],
-    parserOptions: {
-        ecmaVersion: 'latest',
-    },
-};
