+++ backend/.gitignore
@@ -5,3 +5,4 @@ node_modules/
 .env
 .eslintcache
 certs/serviceAccountKey.json
+coverage
