+++ backend/package.json
@@ -4,21 +4,21 @@
   "description": "Backend of VanRoomies project.",
   "main": "server.js",
   "scripts": {
-    "start": "cross-env NODE_ENV=production node src/app.js",
-    "dev": "cross-env NODE_ENV=development nodemon src/app.js",
+    "start": "cross-env NODE_ENV=production node src/server.js",
+    "dev": "cross-env NODE_ENV=development nodemon src/server.js",
     "lint": "npm run prettier",
     "prettier": "prettier --write './src/**/*.js'",
-    "eslint": "eslint --cache --ext .js ./src/"
+    "test": "jest --maxWorkers=50% --coverage=true"
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
     "dotenv": "^16.3.1",
     "express": "^4.18.2",
-    "google-auth-library": "^9.1.0",
-    "jsonwebtoken": "^9.0.2",
     "express-mongo-sanitize": "^2.2.0",
     "firebase-admin": "^11.11.0",
+    "google-auth-library": "^9.1.0",
+    "jsonwebtoken": "^9.0.2",
     "mongoose": "^7.6.0",
     "socket.io": "^4.7.2",
     "validator": "^13.11.0"
@@ -27,6 +27,7 @@
     "concurrently": "^8.2.1",
     "cross-env": "^7.0.3",
     "eslint": "^8.51.0",
+    "jest": "^29.7.0",
     "nodemon": "^3.0.1",
     "prettier": "3.0.3"
   }
