+++ backend/src/app.js
@@ -1,34 +1,9 @@
 const express = require('express');
-const mongoose = require('mongoose');
 const bodyParser = require('body-parser');
-const https = require('https');
-const fs = require('fs');
 const mongoSanitize = require('express-mongo-sanitize');
-const { initializeApp, applicationDefault } = require('firebase-admin/app');
-
 const { logErrors, errorHandler } = require('./middlewares');
-const User = require('./models/userModel');
-
-// Configure environment variable path
-require('dotenv').config({ path: `./.env.${process.env.NODE_ENV}` });
-
-const credentials = {
-    key: fs.readFileSync('certs/key.pem'),
-    cert: fs.readFileSync('certs/cert.pem'),
-};
 
 const app = express();
-const httpsServer = https.createServer(credentials, app);
-const port = process.env.PORT || 3000;
-
-// Connect to MongoDB server through URI from environment variable
-mongoose.connect(process.env.MONGODB_TEST_URI);
-
-const db = mongoose.connection;
-db.on('error', console.error.bind(console, 'Connection error! Make sure MongoDB server is running! '));
-db.once('open', () => {
-    console.log('Connection successful!');
-});
 
 // middleware
 app.use(bodyParser.json());
@@ -68,14 +43,4 @@ app.post('/api/firebase_token', async (req, res) => {
 app.use(logErrors);
 app.use(errorHandler);
 
-// use socket.io
-require('./sock.js')(httpsServer);
-
-// use firebase admin sdk
-initializeApp({
-    credential: applicationDefault(),
-});
-
-httpsServer.listen(port, () => {
-    console.log(`VanRoomies server at https://localhost:${port}`);
-});
+module.exports = app;
